Move SqsActiveJob's class methods
